WhoIsHomeV0.5
this is our clubs alpha release of our who is home project, it is currently only the code snippets. 
this program uses python, nmap, and an arduino to light up a individually addressable LED strip if a device is actively on the network the pi is connected too. 

this is a submission from a tech club, from the embedded systems team
https://sites.google.com/view/valenciatechclub
https://sites.google.com/view/valencia-robotics

this is for submission for the Pi inscrutables contest, I needed a location to host the code. 
Members who worked on this include
-Dylan Poll ( arduino script and Pi setup and documentation) 
https://www.linkedin.com/in/dylan-poll-4a324a1a2/
-Jonathan De La Cruz (python script and documentation.)
https://www.linkedin.com/in/jonathandelacruz96/